+++ apis/python/src/tiledbsoma/dataframe.py
@@ -1,4 +1,4 @@
-from typing import Any, Iterator, Literal, Optional, Sequence, Tuple, TypeVar, Union
+from typing import Any, Iterator, Literal, Optional, Sequence, TypeVar
 
 import numpy as np
 import pandas as pd
@@ -9,24 +9,23 @@ import tiledb
 import tiledbsoma.libtiledbsoma as clib
 
 from . import util, util_arrow
-from .constants import SOMA_JOINID
+from .constants import SOMA_JOINID, SOMA_ROWID
 from .query_condition import QueryCondition  # type: ignore
-from .soma_collection import CollectionBase
+from .collection import CollectionBase
 from .tiledb_array import TileDBArray
 from .types import Ids, ResultOrder
 
 Slice = TypeVar("Slice", bound=Sequence[int])
 
 
-class IndexedDataFrame(TileDBArray):
+class DataFrame(TileDBArray):
     """
     Represents ``obs``, ``var``, and others.
 
-    All ``IndexedDataFrame`` must contain a column called ``soma_joinid``, of type ``int64``. The ``soma_joinid`` column contains a unique value for each row in the ``IndexedDataFrame``, and intended to act as a joint key for other objects, such as ``SparseNdArray``.
+    A ``DataFrame`` contains a "pseudo-column" called ``soma_rowid``, of type int64 and domain [0,num_rows).  The ``soma_rowid`` pseudo-column contains a unique value for each row in the ``DataFrame``, and is intended to act as a join key for other objects, such as a ``SparseNdArray``.
     """
 
-    _index_column_names: Union[Tuple[()], Tuple[str, ...]]
-    _is_sparse: Optional[bool]
+    _cached_is_sparse: Optional[bool]
 
     def __init__(
         self,
@@ -39,27 +38,22 @@ class IndexedDataFrame(TileDBArray):
         See also the ``TileDBOject`` constructor.
         """
         super().__init__(uri=uri, parent=parent, ctx=ctx)
-        self._index_column_names = ()
-        self._is_sparse = None
+        self._cached_is_sparse = None
 
     @property
-    def soma_type(self) -> Literal["IndexedDataFrame"]:
-        return "IndexedDataFrame"
+    def soma_type(self) -> Literal["DataFrame"]:
+        return "DataFrame"
 
     def create(
         self,
         schema: pa.Schema,
-        index_column_names: Sequence[str],
-    ) -> "IndexedDataFrame":
+    ) -> "DataFrame":
         """
-        :param schema: Arrow Schema defining the per-column schema. This schema must define all columns, including columns to be named as index columns. If the schema includes types unsupported by the SOMA implementation, an error will be raised.
-
-        :param index_column_names: A list of column names to use as user-defined index columns (e.g., ``['cell_type', 'tissue_type']``). All named columns must exist in the schema, and at least one index column name is required.
+        :param schema: Arrow Schema defining the per-column schema. This schema must define all columns. The column name ``soma_rowid`` is reserved for the pseudo-column of the same name.  If the schema includes types unsupported by the SOMA implementation, an error will be raised.
         """
-        schema = _validate_schema(schema, index_column_names)
-        self._create_empty(schema, index_column_names)
-        self._is_indexed = True
-        self._index_column_names = tuple(index_column_names)
+        schema = _validate_schema(schema)
+        self._create_empty(schema)
+        self._is_indexed = False
 
         self._common_create()  # object-type metadata etc
         return self
@@ -67,47 +61,26 @@ class IndexedDataFrame(TileDBArray):
     def _create_empty(
         self,
         schema: pa.Schema,
-        index_column_names: Sequence[str],
     ) -> None:
         """
-        Create a TileDB 1D sparse array with dimensions and attributes
+        Create a TileDB 1D dense array with int64 ``soma_rowid`` dimension and multiple attributes.
         """
 
         level = self._tiledb_platform_config.string_dim_zstd_level
 
-        dims = []
-        for index_column_name in index_column_names:
-            dtype = util_arrow.tiledb_type_from_arrow_type(
-                schema.field(index_column_name).type
-            )
-            # We need domain=(None,None) for string dims
-            lo: Any = None
-            hi: Any = None
-            if dtype != str:
-                if np.issubdtype(dtype, np.integer):
-                    lo = np.iinfo(dtype).min
-                    hi = np.iinfo(dtype).max - 1
-                elif np.issubdtype(dtype, np.floating):
-                    lo = np.finfo(dtype).min
-                    hi = np.finfo(dtype).max
-                else:
-                    raise TypeError(f"Unsupported dtype {dtype}")
-            dim = tiledb.Dim(
-                name=index_column_name,
-                domain=(lo, hi),
+        dom = tiledb.Domain(
+            tiledb.Dim(
+                name=SOMA_ROWID,
+                domain=(0, np.iinfo(np.int64).max - 1),
                 tile=2048,  # TODO: PARAMETERIZE
-                dtype=dtype,
+                dtype=np.int64,
                 filters=[tiledb.ZstdFilter(level=level)],
-            )
-            dims.append(dim)
+            ),
+            ctx=self._ctx,
+        )
 
-        dom = tiledb.Domain(dims, ctx=self._ctx)
-
-        attrs = []
-        for attr_name in schema.names:
-            if attr_name in index_column_names:
-                continue
-            attr = tiledb.Attr(
+        attrs = [
+            tiledb.Attr(
                 name=attr_name,
                 dtype=util_arrow.tiledb_type_from_arrow_type(
                     schema.field(attr_name).type
@@ -115,11 +88,15 @@ class IndexedDataFrame(TileDBArray):
                 filters=[tiledb.ZstdFilter()],
                 ctx=self._ctx,
             )
-            attrs.append(attr)
+            for attr_name in schema.names
+            if attr_name != SOMA_ROWID
+        ]
 
         sch = tiledb.ArraySchema(
             domain=dom,
             attrs=attrs,
+            # TODO: pending tiledb issue involving dense dataframes
+            # sparse=False,
             sparse=True,
             allows_duplicates=False,
             offsets_filters=[
@@ -129,13 +106,11 @@ class IndexedDataFrame(TileDBArray):
             ],
             capacity=100000,
             cell_order="row-major",
-            # As of TileDB core 2.8.2, we cannot consolidate string-indexed sparse arrays with
-            # col-major tile order: so we write ``X`` with row-major tile order.
             tile_order="row-major",
             ctx=self._ctx,
         )
-        self._is_sparse = sch.sparse
 
+        self._cached_is_sparse = sch.sparse
         tiledb.Array.create(self._uri, sch, ctx=self._ctx)
 
     def keys(self) -> Sequence[str]:
@@ -145,46 +120,39 @@ class IndexedDataFrame(TileDBArray):
         return self._tiledb_array_keys()
 
     @property
-    def is_indexed(self) -> Literal[True]:
-        return True
+    def is_indexed(self) -> Literal[False]:
+        return False
 
     def get_index_column_names(self) -> Sequence[str]:
-        """
-        Return index (dimension) column names.
-        """
-        # If we've cached the answer, skip the storage read. Especially if the storage is on the
-        # cloud, where we'll avoid an HTTP request.
-        if self._index_column_names == ():
-            assert self.is_indexed
-            self._index_column_names = self._tiledb_dim_names()
-
-        return self._index_column_names
+        return ()
 
     def read(
         self,
         *,
-        # TODO: find out how to spell this in a way the type-checker will accept :(
-        # ids: Optional[Union[Sequence[int], str, Slice]] = None,
+        # TODO: find the right syntax to get the typechecker to accept args like ``ids=slice(0,10)``
+        # ids: Optional[Union[Sequence[int], Slice]] = None,
         ids: Optional[Any] = None,
         value_filter: Optional[str] = None,
         column_names: Optional[Sequence[str]] = None,
         result_order: Optional[ResultOrder] = None,
-        # TODO: more arguments
+        # TODO: batch_size
+        # TODO: partition,
+        # TODO: platform_config,
     ) -> Iterator[pa.Table]:
         """
-        Read a user-defined subset of data, addressed by the dataframe indexing columns, optionally filtered, and return results as one or more Arrow.Table.
+        Read a user-defined subset of data, addressed by the dataframe indexing column, optionally filtered, and return results as one or more ``Arrow.Table``.
 
-        :param ids: for each index dimension, which rows to read. Defaults to ``None``, meaning no constraint -- all IDs.
+        :param ids: Which rows to read. Defaults to ``None``, meaning no constraint -- all rows.
 
         :param column_names: the named columns to read and return. Defaults to ``None``, meaning no constraint -- all column names.
 
         :param partitions: an optional ``ReadPartitions`` hint to indicate how results should be organized.
 
-        :param result_order: order of read results. This can be one of 'row-major', 'col-major', or 'unordered'.
+        :param result_order: order of read results.  This can be one of 'row-major', 'col-major', or 'unordered'.
 
         :param value_filter: an optional [value filter] to apply to the results. Defaults to no filter.
 
-        **Indexing**: the ``ids`` parameter will support, per dimension: a list of values of the type of the indexed column.
+        **Indexing**: the ``ids`` parameter will support, per dimension: a row offset (uint), a row-offset range (slice), or a list of both.
         """
         with self._tiledb_open("r") as A:
             query_condition = None
@@ -200,17 +168,16 @@ class IndexedDataFrame(TileDBArray):
             )
 
             if ids is not None:
-                dim_name = A.schema.domain.dim(0).name
                 if isinstance(ids, list):
-                    sr.set_dim_points(dim_name, ids)
+                    sr.set_dim_points(SOMA_ROWID, ids)
                 elif isinstance(ids, pa.ChunkedArray):
-                    sr.set_dim_points(dim_name, ids)
+                    sr.set_dim_points(SOMA_ROWID, ids)
                 elif isinstance(ids, pa.Array):
-                    sr.set_dim_points(dim_name, pa.chunked_array(ids))
+                    sr.set_dim_points(SOMA_ROWID, pa.chunked_array(ids))
                 elif isinstance(ids, slice):
                     lo_hi = util.slice_to_range(ids)
                     if lo_hi is not None:
-                        sr.set_dim_ranges(dim_name, lo_hi)
+                        sr.set_dim_ranges(SOMA_ROWID, [lo_hi])
 
             # TODO: platform_config
             # TODO: batch_size
@@ -257,10 +224,11 @@ class IndexedDataFrame(TileDBArray):
         result_order: Optional[ResultOrder] = None,
         # TODO: batch_size
         # TODO: partition,
+        # TODO: result_order,
         # TODO: platform_config,
     ) -> pa.Table:
         """
-        This is a convenience method around ``read``. It iterates the return value from ``read`` and returns a concatenation of all the table-pieces found. Its nominal use is to simply unit-test cases.
+        This is a convenience method around ``read``. It concatenates all partial read results into a single Table. Its nominal use is to simplify unit-test cases.
         """
         return pa.concat_tables(
             self.read(
@@ -271,39 +239,73 @@ class IndexedDataFrame(TileDBArray):
             )
         )
 
+    def _get_is_sparse(self) -> bool:
+        if self._cached_is_sparse is None:
+
+            # Simpler would be:
+            # if self.exists():
+            #     with self._tiledb_open("r") as A:
+            #         self._cached_is_sparse = A.schema.sparse
+            # but that has _two_ HTTP round trips in the tiledb-cloud case.
+            # This way, there is only one.
+            try:
+                with self._tiledb_open("r") as A:
+                    self._cached_is_sparse = A.schema.sparse
+            except tiledb.TileDBError as e:
+                raise Exception(f"could not read array schema at {self._uri}") from e
+
+        return self._cached_is_sparse
+
     def write(self, values: pa.Table) -> None:
         """
-        Write an Arrow.Table to the persistent object. As duplicate index values are not allowed, index values already present in the object are overwritten and new index values are added.
+        Write an Arrow.Table to the persistent object.
 
-        :param values: An Arrow.Table containing all columns, including the index columns. The schema for the values must match the schema for the ``IndexedDataFrame``.
+        :param values: An Arrow.Table containing all columns, including the index columns. The schema for the values must match the schema for the ``DataFrame``.
+
+        The ``values`` Arrow Table must contain a ``soma_rowid`` (int64) column, indicating which rows are being written.
         """
-        dim_cols_list = []
+        if (
+            SOMA_ROWID not in values.schema.names
+            or SOMA_JOINID not in values.schema.names
+        ):
+            raise ValueError(
+                f"Write requires both {SOMA_ROWID} and {SOMA_JOINID} to be specified in table."
+            )
+
+        # TODO: contiguity check, and/or split into multiple contiguous writes
+        # For now: just assert that these _already are_ contiguous and start with 0.
+        rowids = [e.as_py() for e in values.column(SOMA_ROWID)]
+
         attr_cols_map = {}
-        dim_names_set = self.get_index_column_names()
-        n = None
-
         for name in values.schema.names:
-            n = len(values.column(name))
-            if name in dim_names_set:
-                dim_cols_list.append(values.column(name).to_pandas())
-            else:
-                attr_cols_map[name] = values.column(name).to_pandas()
-        assert n is not None
+            if name != SOMA_ROWID:
+                attr_cols_map[name] = np.asarray(
+                    values.column(name).to_pandas(
+                        types_mapper=util_arrow.tiledb_type_from_arrow_type_for_write,
+                    )
+                )
 
-        dim_cols_list = [list(dim_col) for dim_col in dim_cols_list]
-        with self._tiledb_open("w") as A:
-            # TODO: find the right syntax for vardims ... it's not the ``*`` operator ...
-            # A[*dim_cols_list] = attr_cols_map
-            if len(dim_cols_list) == 1:
-                A[dim_cols_list[0]] = attr_cols_map
-            elif len(dim_cols_list) == 2:
-                A[dim_cols_list[0], dim_cols_list[1]] = attr_cols_map
-            else:
-                raise Exception("ndim >= 2 not currently supported")
+        if self._get_is_sparse():
+            # sparse write
+            with self._tiledb_open("w") as A:
+                A[rowids] = attr_cols_map
+        else:
+            # TODO: This was a quick thing to bootstrap some early ingestion tests but needs more thought.
+            # In particular, rowids needn't be either zero-up or contiguous.
+            assert len(rowids) > 0
+            rowids = sorted(rowids)
+            assert rowids[0] == 0
+
+            # dense write
+            lo = rowids[0]
+            hi = rowids[-1]
+            with self._tiledb_open("w") as A:
+                A[lo : (hi + 1)] = attr_cols_map
 
     def read_as_pandas(
         self,
-        ids: Optional[Any] = None,
+        *,
+        ids: Optional[Ids] = None,
         value_filter: Optional[str] = None,
         column_names: Optional[Sequence[str]] = None,
         result_order: Optional[ResultOrder] = None,
@@ -318,6 +320,7 @@ class IndexedDataFrame(TileDBArray):
 
     def read_as_pandas_all(
         self,
+        *,
         ids: Optional[Ids] = None,
         value_filter: Optional[str] = None,
         column_names: Optional[Sequence[str]] = None,
@@ -333,44 +336,41 @@ class IndexedDataFrame(TileDBArray):
             ignore_index=True,
         )
 
-    def write_from_pandas(
-        self,
-        dataframe: pd.DataFrame,
-    ) -> None:
+    def write_from_pandas(self, dataframe: pd.DataFrame) -> None:
+        """
+        Write the Pandas DataFrame. The Pandas DataFrame must contain a soma_rowid and soma_joinid
+        column of type int64.
+        """
         self.write(pa.Table.from_pandas(dataframe))
 
 
-def _validate_schema(schema: pa.Schema, index_column_names: Sequence[str]) -> pa.Schema:
+def _validate_schema(schema: pa.Schema) -> pa.Schema:
     """
-    Handle default column additions (eg, soma_joinid) and error checking on required columns.
+    Handle default column additions (eg, soma_rowid) and error checking on required columns.
 
     Returns a schema, which may be modified by the addition of required columns.
     """
-    if not index_column_names:
-        raise ValueError("IndexedDataFrame requires one or more index columns")
+    if SOMA_ROWID in schema.names:
+        if schema.field(SOMA_ROWID).type != pa.int64():
+            raise TypeError(f"{SOMA_ROWID} field must be of type Arrow int64")
+    else:
+        # add SOMA_ROWID
+        schema = schema.insert(0, pa.field(SOMA_ROWID, pa.int64()))
 
     if SOMA_JOINID in schema.names:
         if schema.field(SOMA_JOINID).type != pa.int64():
-            raise ValueError(f"{SOMA_JOINID} field must be of type Arrow int64")
+            raise TypeError(f"{SOMA_JOINID} field must be of type Arrow int64")
     else:
         # add SOMA_JOINID
-        schema = schema.append(pa.field(SOMA_JOINID, pa.int64()))
+        schema = schema.insert(1, pa.field(SOMA_JOINID, pa.int64()))
 
-    # verify no illegal use of soma_ prefix
     for field_name in schema.names:
-        if field_name.startswith("soma_") and field_name != SOMA_JOINID:
+        if field_name.startswith("soma_") and field_name not in [
+            SOMA_ROWID,
+            SOMA_JOINID,
+        ]:
             raise ValueError(
-                "IndexedDataFrame schema may not contain fields with name prefix `soma_`"
+                "DataFrame schema may not contain fields with name prefix `soma_`"
             )
-
-    # verify that all index_column_names are present in the schema
-    schema_names_set = set(schema.names)
-    for index_column_name in index_column_names:
-        if index_column_name.startswith("soma_") and index_column_name != SOMA_JOINID:
-            raise ValueError(
-                "IndexedDataFrame schema may not contain fields with name prefix `soma_`"
-            )
-        if index_column_name not in schema_names_set:
-            raise ValueError("All index names must be dataframe schema")
 
     return schema
